Crate importer

It’s assumed that the importer is run with the current directory as the project repository; aka ./data/ and ./importer/config must exist.

Modules

berlin 🔒

configuration 🔒

map_config 🔒

pick_geofabrik 🔒

seattle 🔒

soundcast 🔒

uk 🔒

utils 🔒

Structs

ImporterConfiguration

Job

A specification for importing all maps in a single city.

Functions

oneshot

Transforms a .osm.xml or .pbf file to a map in one step.

osmium

Uses osmium to clip the input .osm.xml or osm.pbf against a polygon and produce some output pbf file. Skips if the output exists.

pick_geofabrik

Given the path to a GeoJSON boundary polygon, return the URL of the smallest Geofabrik osm.pbf file that completely covers the boundary, and the path to where the local copy should go.

regenerate_everything

Regenerate all maps and scenarios from scratch.