Module importer::actdev

This module is used for all cities associated with the https://github.com/cyipt/actdev project.

Structs

ODSummary
Describes some number of people that have the same home, workplace, and preferred mode. When they're created, the only thing that'll differ between them is exact departure time.

Functions

generate_scenario

import_scenarios

parse_usize

rand_duration

rand_time