Struct importer::generic::GenericCityImporter

pub struct GenericCityImporter {
    pub osm_url: String,
    pub map_config: MapConfig,
    pub onstreet_parking: OnstreetParking,
    pub public_offstreet_parking: PublicOffstreetParking,
    pub private_offstreet_parking: PrivateOffstreetParking,
    pub elevation: Option<String>,
    pub include_railroads: bool,
}

Importing a new city can be done just by filling out this config file and specifying some polygon boundaries. Most fields are directly from convert_osm::Options.

If any extra data is imported for a city (like collisions or population), then for now, don't use this.

Fields
osm_url: String

The URL to a .osm or .osm.pbf file containing the entire city. http://download.geofabrik.de/ is recommended.

map_config: MapConfigonstreet_parking: OnstreetParkingpublic_offstreet_parking: PublicOffstreetParkingprivate_offstreet_parking: PrivateOffstreetParkingelevation: Option<String>

If provided, pull elevation data from this SRTM file. The SRTM parser is incorrect, so the results will be nonsense.

include_railroads: bool

OSM railway=rail will be included as light rail if so. Cosmetic only.

Implementations

impl GenericCityImporter

pub fn osm_to_raw(
&self,
name: MapName,
timer: &mut Timer<'_>,
config: &ImporterConfiguration
) -> RawMap