Function importer::map_config::config_for_map

pub fn config_for_map(name: &MapName) -> Options

Given the name of a map, configure its import.

Note this was once expressed as config files for every city. That was less maintainable; most places used default values that were copied around.