importer
Module pick_geofabrik

Functions

find_matching_regions 🔒

load_boundary 🔒

load_remote_geojson 🔒

pick_geofabrik

Given the path to a GeoJSON boundary polygon, return the URL of the smallest Geofabrik osm.pbf file that completely covers the boundary, and the path to where the local copy should go.