Crate importer

It’s assumed that the importer is run with the current directory as the project repository; aka ./data/ and ./importer/config must exist.

Modules

basemap 🔒

berlin 🔒

configuration 🔒

map_config 🔒

pick_geofabrik 🔒

seattle 🔒

soundcast 🔒

uk 🔒

utils 🔒

Structs

Job

A specification for importing all maps in a single city.

Functions

oneshot

Transforms a .osm file to a map in one step.

pick_geofabrik

Given the path to an Osmosis boundary polygon, return the URL of a Geofabrik osm.pbf file that best covers the boundary.

regenerate_everything

Regenerate all maps and scenarios from scratch.