Function importer::soundcast::trips::endpoints

fn endpoints(
    from: &Endpoint, 
    to: &Endpoint, 
    map: &Map, 
    osm_id_to_bldg: &HashMap<OsmID, BuildingID>, 
    (in_borders, out_borders): (&Vec<(IntersectionID, LonLat)>, &Vec<(IntersectionID, LonLat)>), 
    constraints: PathConstraints, 
    maybe_huge_map: Option<&(&Map, HashMap<OsmID, BuildingID>)>, 
    only_passthrough_trips: bool
) -> Option<(TripEndpoint, TripEndpoint)>

Transform the Seattle-wide Endpoints into specific TripEndpoints for this map. When the endpoint happens to be a building on the map, this is straightforward. Otherwise, the endpoint will snap to a border intersection.

When only_passthrough_trips is true, only trips beginning and ending off-map are returned. When it’s false, all other trips are returned.